requirements.txt
groq==0.4.2
qdrant-client==2.7.0
langchain==0.1.5
langchain-groq==0.0.1
langchain-community==0.0.10
langchain-core==0.1.19
PyPDF2==3.0.1
numpy==1.24.3
sentence-transformers==2.2.2
transformers==4.30.2
gradio==4.0.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
langchainhub==0.1.10
python-dotenv==1.0.0